FaceAccessories/requirements.txt
streamlit>=1.32.0
opencv-python-headless>=4.9.0
mediapipe>=0.10.9
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.2.0
matplotlib>=3.8.0
